Read me HOK is Hands Of Karma
The idea is not everyone knows where to start life 
 many people have a rough start at life. That's when drugs and crime take over and rule or rule over those who havnt had the education to combat  the negative with with positive. 
I want to start a project, profit or nonprofit that hands on teaches everything. From the worst of the worst  addicts to the people fresh out of high school  clean and freash and gear or the kids in-between. All helping with a hand up and paying it forward and getting the same from some that
From coding apps to programming  there own smart phones or computers to, doing the same with homes and rental property managment to owning their own earn I no everything as they grow their education. Anything imaginable could grow from this and will . That could all fit under the umbrella of (a better world), (an educated  society), (a caring environment)  Karma  is  bad it's equally good. It's a get out what you put in kind of idea evolving all positive for a positive!!!  It's all good intentions so it can't fail. Maybe an app to kick it off locally or on a local level. I have all home and real-estate  knowledge construction and legal. Rental property's I own and manage to managing other people's. I'll put all I have  all I know  and as much work into this as I can. Who's in?
- 📫 How to reach me at. 918-934-4757
                          itsabluworld@gmail.com 
                        Itsabluworld@hotmail.com
                            Itsabluworld@aol.com 
                        Itsabluworld@outlook.com
                           or any other mail.com
                               or any other mail.com such as yahoo.com im available at anytime. Feel free to call or text my office cell. 918-934-4757 or come on down to Inola Oklahoma and ask around they all know me down here at the Hey capital of the world.